+++ src/app/common/pipes/display-dataset-id.pipe.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from "@angular/core";
+import { CommonModule } from "@angular/common";
+import { DisplayDatasetIdPipe } from "./display-dataset-id.pipe";
+
+@NgModule({
+    declarations: [DisplayDatasetIdPipe],
+    exports: [DisplayDatasetIdPipe],
+    imports: [CommonModule],
+})
+export class DisplayDatasetIdPipeModule {}