+++ src/app/components/dataset-list-component/dataset-list.module.ts
@@ -12,6 +12,7 @@ import { DatasetListComponent } from "./dataset-list.component";
 import { MatSelectModule } from "@angular/material/select";
 import { AngularSvgIconModule } from "angular-svg-icon";
 import { RouterModule } from "@angular/router";
+import { DatasetVisibilityModule } from "../dataset-visibility/dataset-visibility.module";
 import { SharedModule } from "src/app/shared/shared/shared.module";
 
 @NgModule({
@@ -27,6 +28,7 @@ import { SharedModule } from "src/app/shared/shared/shared.module";
         AngularSvgIconModule,
         DisplayTimeModule,
         RouterModule,
+        DatasetVisibilityModule,
         SharedModule,
     ],
     exports: [DatasetListComponent, DatasetListItemComponent],