+++ src/app/components/dataset-visibility/dataset-visibility.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from "@angular/core";
+import { CommonModule } from "@angular/common";
+import { DatasetVisibilityComponent } from "./dataset-visibility.component";
+
+@NgModule({
+    declarations: [DatasetVisibilityComponent],
+    exports: [DatasetVisibilityComponent],
+    imports: [CommonModule],
+})
+export class DatasetVisibilityModule {}