+++ src/app/components/lineage-graph/lineage-graph.component.ts
@@ -18,6 +18,7 @@ import AppValues from "src/app/common/app.values";
 import {
     LineageGraph,
     LineageGraphNodeKind,
+    LineageNodeAccess,
 } from "src/app/dataset-view/additional-components/lineage-component/lineage-model";
 import { LineageGraphConfig, LINEAGE_CONFIG } from "./ligeage-graph.settings";
 import { SessionStorageService } from "src/app/services/session-storage.service";
@@ -33,7 +34,9 @@ export class LineageGraphComponent implements OnInit, OnChanges {
     @Input({ required: true }) public graph: LineageGraph;
     @Input({ required: true }) public currentDataset: DatasetLineageBasicsFragment;
     @Output() public onClickNodeEvent = new EventEmitter<Node>();
+    @Output() public onClickPrivateNodeEvent = new EventEmitter<Node>();
     public readonly LineageGraphNodeKind: typeof LineageGraphNodeKind = LineageGraphNodeKind;
+    public readonly LineageNodeAccess: typeof LineageNodeAccess = LineageNodeAccess;
     public readonly DatasetKind: typeof DatasetKind = DatasetKind;
     public readonly LINEAGE_CONFIG: LineageGraphConfig = LINEAGE_CONFIG;
     public readonly DEFAULT_AVATAR_URL = AppValues.DEFAULT_AVATAR_URL;
@@ -76,9 +79,17 @@ export class LineageGraphComponent implements OnInit, OnChanges {
         this.onClickNodeEvent.emit(node);
     }
 
+    public onClickPrivateNode(node: Node): void {
+        this.onClickPrivateNodeEvent.emit(node);
+    }
+
     public onClickInfo(): void {
         this.showSidePanel = !this.showSidePanel;
         this.sessionStorageService.setSidePanelVisible(this.showSidePanel);
+    }
+
+    public get isPrivateVisibility(): boolean {
+        return this.currentDataset.visibility.__typename === "PrivateDatasetVisibility";
     }
 
     private checkVisibilitySidePanel(): void {