+++ src/app/dataset-view/additional-components/data-tabs.mock.ts
@@ -11,8 +11,9 @@ import {
     LineageUpdate,
     MetadataSchemaUpdate,
 } from "../dataset.subscriptions.interface";
-import { mockDatasetBasicsDerivedFragment } from "src/app/search/mock.data";
+import { mockDatasetBasicsDerivedFragment, mockPublicDatasetVisibility } from "src/app/search/mock.data";
 import { Node } from "@swimlane/ngx-graph";
+import { LineageNodeAccess } from "./lineage-component/lineage-model";
 import { OperationColumnClassEnum } from "src/app/interface/dataset.interface";
 
 export const mockDataUpdate: DataUpdate = {
@@ -155,6 +156,7 @@ export const mockLineageGraphUpdate: LineageUpdate = {
             avatarUrl: "https://avatars.githubusercontent.com/u/50896974?s=200&v=4",
         },
         alias: "kamu/alberta.case-details",
+        visibility: mockPublicDatasetVisibility,
         metadata: {
             __typename: "DatasetMetadata",
             currentPollingSource: {
@@ -195,6 +197,8 @@ export const mockLineageGraphUpdate: LineageUpdate = {
                 avatarUrl: "https://avatars.githubusercontent.com/u/50896974?s=200&v=4",
             },
             alias: "kamu/alberta.case-details",
+            visibility: mockPublicDatasetVisibility,
+
             metadata: {
                 __typename: "DatasetMetadata",
                 currentPollingSource: {
@@ -252,6 +256,7 @@ export const mockLineageGraphUpdate: LineageUpdate = {
             kind: DatasetKind.Derivative,
             name: "alberta.case-details.hm",
             alias: "alberta.case-details.hm",
+            visibility: mockPublicDatasetVisibility,
         },
         {
             __typename: "Dataset",
@@ -284,6 +289,7 @@ export const mockLineageGraphUpdate: LineageUpdate = {
             kind: DatasetKind.Derivative,
             name: "canada.case-details",
             alias: "canada.case-details",
+            visibility: mockPublicDatasetVisibility,
         },
         {
             __typename: "Dataset",
@@ -316,6 +322,7 @@ export const mockLineageGraphUpdate: LineageUpdate = {
             kind: DatasetKind.Derivative,
             name: "canada.daily-cases",
             alias: "canada.daily-cases",
+            visibility: mockPublicDatasetVisibility,
         },
     ],
     edges: [
@@ -332,6 +339,7 @@ export const mockLineageGraphUpdate: LineageUpdate = {
                     avatarUrl: "https://avatars.githubusercontent.com/u/50896974?s=200&v=4",
                 },
                 alias: "kamu/alberta.case-details",
+                visibility: mockPublicDatasetVisibility,
                 metadata: {
                     __typename: "DatasetMetadata",
                     currentPollingSource: {
@@ -389,6 +397,7 @@ export const mockLineageGraphUpdate: LineageUpdate = {
                 kind: DatasetKind.Derivative,
                 name: "alberta.case-details.hm",
                 alias: "alberta.case-details.hm",
+                visibility: mockPublicDatasetVisibility,
             },
         ],
         [
@@ -423,6 +432,7 @@ export const mockLineageGraphUpdate: LineageUpdate = {
                 kind: DatasetKind.Derivative,
                 name: "alberta.case-details.hm",
                 alias: "alberta.case-details.hm",
+                visibility: mockPublicDatasetVisibility,
             },
             {
                 __typename: "Dataset",
@@ -455,6 +465,7 @@ export const mockLineageGraphUpdate: LineageUpdate = {
                 kind: DatasetKind.Derivative,
                 name: "canada.case-details",
                 alias: "canada.case-details",
+                visibility: mockPublicDatasetVisibility,
             },
         ],
         [
@@ -489,6 +500,7 @@ export const mockLineageGraphUpdate: LineageUpdate = {
                 kind: DatasetKind.Derivative,
                 name: "canada.case-details",
                 alias: "canada.case-details",
+                visibility: mockPublicDatasetVisibility,
             },
             {
                 __typename: "Dataset",
@@ -521,6 +533,7 @@ export const mockLineageGraphUpdate: LineageUpdate = {
                 kind: DatasetKind.Derivative,
                 name: "canada.daily-cases",
                 alias: "canada.daily-cases",
+                visibility: mockPublicDatasetVisibility,
             },
         ],
     ],
@@ -539,6 +552,7 @@ export const mockLineageGraphUpdateWithMqttSource: LineageUpdate = {
             avatarUrl: "https://avatars.githubusercontent.com/u/50896974?s=200&v=4",
         },
         alias: "temp",
+        visibility: mockPublicDatasetVisibility,
         metadata: {
             __typename: "DatasetMetadata",
             currentPollingSource: {
@@ -573,6 +587,7 @@ export const mockLineageGraphUpdateWithMqttSource: LineageUpdate = {
                 avatarUrl: "https://avatars.githubusercontent.com/u/50896974?s=200&v=4",
             },
             alias: "temp",
+            visibility: mockPublicDatasetVisibility,
             metadata: {
                 __typename: "DatasetMetadata",
                 currentPollingSource: {
@@ -617,7 +632,7 @@ export const mockBuildGraphNodesResult: Node[] = [
                 name: "alberta.case-details",
                 kind: "ROOT",
                 isCurrent: true,
-                access: "private",
+                access: LineageNodeAccess.PUBLIC,
                 accountName: "kamu",
                 avatarUrl: "https://avatars.githubusercontent.com/u/50896974?s=200&v=4",
             },
@@ -633,7 +648,7 @@ export const mockBuildGraphNodesResult: Node[] = [
                 name: "alberta.case-details.hm",
                 kind: "DERIVATIVE",
                 isCurrent: false,
-                access: "private",
+                access: LineageNodeAccess.PUBLIC,
                 accountName: "kamu",
                 avatarUrl: "https://avatars.githubusercontent.com/u/50896974?s=200&v=4",
             },
@@ -649,7 +664,7 @@ export const mockBuildGraphNodesResult: Node[] = [
                 name: "canada.case-details",
                 kind: "DERIVATIVE",
                 isCurrent: false,
-                access: "private",
+                access: LineageNodeAccess.PUBLIC,
                 accountName: "kamu",
                 avatarUrl: "https://avatars.githubusercontent.com/u/50896974?s=200&v=4",
             },
@@ -665,7 +680,7 @@ export const mockBuildGraphNodesResult: Node[] = [
                 name: "canada.daily-cases",
                 kind: "DERIVATIVE",
                 isCurrent: false,
-                access: "private",
+                access: LineageNodeAccess.PUBLIC,
                 accountName: "kamu",
                 avatarUrl: "https://avatars.githubusercontent.com/u/50896974?s=200&v=4",
             },
@@ -1467,6 +1482,7 @@ export const mockMetadataDerivedUpdate: MetadataSchemaUpdate = {
                                 accountName: "kamu",
                             },
                             alias: "kamu/alberta.case-details",
+                            visibility: mockPublicDatasetVisibility,
                         },
                     },
                 ],