+++ src/app/dataset-view/dataset.component.spec.ts
@@ -40,7 +40,7 @@ import { HistoryComponent } from "./additional-components/history-component/hist
 import { LineageComponent } from "./additional-components/lineage-component/lineage.component";
 import { DatasetSettingsGeneralTabComponent } from "./additional-components/dataset-settings-component/tabs/general/dataset-settings-general-tab.component";
 import { DatasetSettingsSchedulingTabComponent } from "./additional-components/dataset-settings-component/tabs/scheduling/dataset-settings-scheduling-tab.component";
-import { ToastrModule } from "ngx-toastr";
+import { ToastrModule, ToastrService } from "ngx-toastr";
 import { DataAccessPanelModule } from "../components/data-access-panel/data-access-panel.module";
 import { SqlEditorComponent } from "../shared/editor/components/sql-editor/sql-editor.component";
 import { RequestTimerComponent } from "../query/shared/request-timer/request-timer.component";
@@ -48,6 +48,7 @@ import { EditorModule } from "../shared/editor/editor.module";
 import { MatProgressBarModule } from "@angular/material/progress-bar";
 import { CdkAccordionModule } from "@angular/cdk/accordion";
 import { FlowsComponent } from "./additional-components/flows-component/flows.component";
+import { DatasetVisibilityModule } from "../components/dataset-visibility/dataset-visibility.module";
 import { RouterTestingModule } from "@angular/router/testing";
 import { promiseWithCatch } from "../common/app.helpers";
 import { QueryAndResultSectionsComponent } from "../query/shared/query-and-result-sections/query-and-result-sections.component";
@@ -56,6 +57,7 @@ import { SqlQueryService } from "../services/sql-query.service";
 import { SearchAndSchemasSectionComponent } from "../query/global-query/search-and-schemas-section/search-and-schemas-section.component";
 import { SharedModule } from "../shared/shared/shared.module";
 import { DatasetRequestBySql } from "../interface/dataset.interface";
+import { MOCK_NODES } from "../api/mock/dataset.mock";
 
 describe("DatasetComponent", () => {
     let component: DatasetComponent;
@@ -66,6 +68,7 @@ describe("DatasetComponent", () => {
     let sqlQueryService: SqlQueryService;
     let route: ActivatedRoute;
     let router: Router;
+    let toastrService: ToastrService;
     const MOCK_DATASET_ROUTE = "kamu/mockNameDerived";
 
     beforeEach(async () => {
@@ -110,6 +113,7 @@ describe("DatasetComponent", () => {
                 EditorModule,
                 MatProgressBarModule,
                 CdkAccordionModule,
+                DatasetVisibilityModule,
                 SharedModule,
                 RouterTestingModule.withRoutes([{ path: MOCK_DATASET_ROUTE, component: DatasetComponent }]),
             ],
@@ -170,7 +174,7 @@ describe("DatasetComponent", () => {
         fixture = TestBed.createComponent(DatasetComponent);
         router.initialNavigation();
         route = TestBed.inject(ActivatedRoute);
-
+        toastrService = TestBed.inject(ToastrService);
         navigationService = TestBed.inject(NavigationService);
         component = fixture.componentInstance;
         component.datasetBasics = mockDatasetBasicsDerivedFragment;
@@ -354,5 +358,11 @@ describe("DatasetComponent", () => {
 
         component.ngOnDestroy();
         expect(removeDatasetSqlCodeSpy).toHaveBeenCalledTimes(1);
+    });
+
+    it("should check click on private node", () => {
+        const toastrServiceSuccessSpy = spyOn(toastrService, "success");
+        component.onClickPrivateLineageNode(MOCK_NODES[2]);
+        expect(toastrServiceSuccessSpy).toHaveBeenCalledWith("Copied ID");
     });
 });