+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Create dataset: unlock dataset visibility selection
 - Saved sql request in the URL after completing the network request
 ### Changed
 - Environment variables and secrets: made `key` field unavailable for editing